move connection tests to modules
